Fix `Ui::Animations::Simple`'s `.start()` and `.change()` methods
Unlike the `.change()` method, which configures a running animation to
go to a different state, the `.start()` method should first set the
initial opacity value, then animate.

This commit fixes this behavior. It changes the private `.prepare()`
method to interpret negative `from` values as "unchanged from" (only
available if the animation is already running), and fixes its usage in
`.change()`.